Fix dynamic code generation.
Previously, dynamic code generation was not scoped, so re-generating a library
using the same source file would re-use the old library. This would cause bad
DFT code to not be tested if it was generated after an initial good DFT.

Now the use of a dynamically-loaded library is scoped, so it will automatically
be closed.